+++ crates/ruma-events/src/room.rs
@@ -10,6 +10,7 @@ use ruma_common::{
     OwnedMxcUri,
 };
 use serde::{de, Deserialize, Serialize};
+use zeroize::{Zeroize, ZeroizeOnDrop};
 
 pub mod aliases;
 pub mod avatar;
@@ -223,22 +224,25 @@ impl From<EncryptedFileInit> for EncryptedFile {
 ///
 /// To create an instance of this type, first create a `JsonWebKeyInit` and convert it via
 /// `JsonWebKey::from` / `.into()`.
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Deserialize, Serialize, Zeroize, ZeroizeOnDrop)]
 #[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
 pub struct JsonWebKey {
     /// Key type.
     ///
     /// Must be `oct`.
+    #[zeroize(skip)]
     pub kty: String,
 
     /// Key operations.
     ///
     /// Must at least contain `encrypt` and `decrypt`.
+    #[zeroize(skip)]
     pub key_ops: Vec<String>,
 
     /// Algorithm.
     ///
     /// Must be `A256CTR`.
+    #[zeroize(skip)]
     pub alg: String,
 
     /// The key, encoded as url-safe unpadded base64.
@@ -248,6 +252,7 @@ pub struct JsonWebKey {
     ///
     /// Must be `true`. This is a
     /// [W3C extension](https://w3c.github.io/webcrypto/#iana-section-jwk).
+    #[zeroize(skip)]
     pub ext: bool,
 }